FU-79 refactor: product 관련 클래스 각 도메인 별 패키지 구조로 분리
