--- src/main/java/com/foru/freebe/product/dto/ProductOptionDto.java
@@ -1,12 +1,0 @@
-package com.foru.freebe.product.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class ProductOptionDto {
-	private String title;
-	private Integer price;
-	private String description;
-}
