--- src/main/java/com/foru/freebe/product/entity/DiscountType.java
@@ -1,6 +1,0 @@
-package com.foru.freebe.product.entity;
-
-public enum DiscountType {
-	RATE,
-	AMOUNT
-}
