+++ src/main/java/com/foru/freebe/productComponent/dto/ProductComponentDto.java
@@ -1,0 +1,12 @@
+package com.foru.freebe.productComponent.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class ProductComponentDto {
+    private String title;
+    private String content;
+    private String description;
+}
