+++ src/main/java/com/foru/freebe/productComponent/service/ProductComponentService.java
@@ -1,9 +1,9 @@
-package com.foru.freebe.product.service;
+package com.foru.freebe.productComponent.service;
 
 import java.util.List;
 
-import com.foru.freebe.product.dto.ProductComponentDto;
 import com.foru.freebe.product.entity.Product;
+import com.foru.freebe.productComponent.dto.ProductComponentDto;
 
 public interface ProductComponentService {
     void registerProductComponent(List<ProductComponentDto> productComponentDtoList, Product product);
