+++ src/main/java/com/foru/freebe/productDiscount/entity/DiscountType.java
@@ -1,0 +1,6 @@
+package com.foru.freebe.productDiscount.entity;
+
+public enum DiscountType {
+    RATE,
+    AMOUNT
+}
