+++ src/main/java/com/foru/freebe/productDiscount/repository/ProductDiscountRepository.java
@@ -1,8 +1,8 @@
-package com.foru.freebe.product.respository;
+package com.foru.freebe.productDiscount.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.foru.freebe.product.entity.ProductDiscount;
+import com.foru.freebe.productDiscount.entity.ProductDiscount;
 
 public interface ProductDiscountRepository extends JpaRepository<ProductDiscount, Long> {
 }
