+++ src/main/java/com/foru/freebe/productDiscount/service/ProductDiscountService.java
@@ -1,9 +1,9 @@
-package com.foru.freebe.product.service;
+package com.foru.freebe.productDiscount.service;
 
 import java.util.List;
 
-import com.foru.freebe.product.dto.ProductDiscountDto;
 import com.foru.freebe.product.entity.Product;
+import com.foru.freebe.productDiscount.dto.ProductDiscountDto;
 
 public interface ProductDiscountService {
     void registerDiscount(List<ProductDiscountDto> productDiscountDtoList, Product product);
