+++ src/main/java/com/foru/freebe/productOption/service/ProductOptionServiceImpl.java
@@ -1,13 +1,13 @@
-package com.foru.freebe.product.service;
+package com.foru.freebe.productOption.service;
 
 import java.util.List;
 
 import org.springframework.stereotype.Service;
 
-import com.foru.freebe.product.dto.ProductOptionDto;
 import com.foru.freebe.product.entity.Product;
-import com.foru.freebe.product.entity.ProductOption;
-import com.foru.freebe.product.respository.ProductOptionRepository;
+import com.foru.freebe.productOption.dto.ProductOptionDto;
+import com.foru.freebe.productOption.entity.ProductOption;
+import com.foru.freebe.productOption.repository.ProductOptionRepository;
 
 import lombok.RequiredArgsConstructor;
